ADD: working merged dev babelless webpack dev configs.
